Moved the code that retrieves the written messages to the DisplayAllMessagesActivity
